+++ .envrc
@@ -1,5 +1,5 @@
 layout python3
 
-export QUART_APP=snikket_web:app
+export QUART_APP='snikket_web:create_app()'
 export QUART_ENV=development
 export SNIKKET_WEB_CONFIG="$(pwd)/.local/web_config.py"
