+++ mypy.ini
@@ -6,13 +6,13 @@ disallow_untyped_calls = True
 disallow_untyped_defs = True
 disallow_incomplete_defs = True
 #check_untyped_defs = True
-disallow_untyped_decorators = False
+disallow_untyped_decorators = True
 #disallow_any_unimported = True
 #disallow_any_expr = True
 #disallow_any_decorated = True
 disallow_any_explicit = False
 #disallow_any_generics = True
-disallow_subclassing_any = False
+disallow_subclassing_any = True
 no_implicit_optional = True
 warn_redundant_casts = True
 warn_unused_ignores = True
