+++ snikket_web/templates/about.html
@@ -19,7 +19,7 @@ Snikket Web Portal ({{ version }})</pre>
 AVATAR_CACHE_TTL = {{ config.get("AVATAR_CACHE_TTL") | repr }}
 SECRET_KEY = &lt;hidden&gt;
 PROSODY_ENDPOINT = &lt;hidden&gt;</pre> #}
-				<p><a href="{{ url_for('home') }}" class="button primary">Back to main page</a></pa>
+				<p><a href="{{ url_for('.home') }}" class="button primary">Back to main page</a></pa>
 			</div>
 		</main>
 	</body>
