+++ snikket_web/templates/user_passwd.html
@@ -36,7 +36,7 @@
 		<p>{% trans %}After changing your password, you will have to enter the new password on all of your devices.{% endtrans %}</p>
 	</div>
 	<div class="f-bbox">
-		<a href="{{ url_for('user.index') }}" class="button secondary">{% trans %}Back{% endtrans %}</a>
+		<a href="{{ url_for('.index') }}" class="button secondary">{% trans %}Back{% endtrans %}</a>
 		<button type="submit" class="primary">{% trans %}Change password{% endtrans %}</button>
 	</div>
 </form></div>
