+++ snikket_web/user.py
@@ -1,24 +1,26 @@
 import typing
 
 import quart.flask_patch
-
 from quart import Blueprint, render_template, request, redirect, url_for
 import quart.exceptions
-from flask_wtf import FlaskForm
-from flask_babel import lazy_gettext as _l, _
+
 import wtforms
 
-from snikket_web.prosodyclient import client
+import flask_wtf
 
-user_bp = Blueprint('user', __name__, url_prefix="/user")
+from flask_babel import lazy_gettext as _l, _
+
+from .infra import client
+
+bp = Blueprint('user', __name__)
 
 
-@user_bp.context_processor
+@bp.context_processor
 async def proc() -> typing.Mapping[str, typing.Any]:
     return {"user_info": await client.get_user_info()}
 
 
-class ChangePasswordForm(FlaskForm):
+class ChangePasswordForm(flask_wtf.FlaskForm):  # type:ignore
     current_password = wtforms.PasswordField(
         _l("Current password"),
         validators=[wtforms.validators.InputRequired()]
@@ -39,11 +41,11 @@ class ChangePasswordForm(FlaskForm):
     )
 
 
-class LogoutForm(FlaskForm):
+class LogoutForm(flask_wtf.FlaskForm):  # type:ignore
     pass
 
 
-class ProfileForm(FlaskForm):
+class ProfileForm(flask_wtf.FlaskForm):  # type:ignore
     nickname = wtforms.TextField(
         _l("Display name"),
     )
@@ -53,14 +55,14 @@ class ProfileForm(FlaskForm):
     )
 
 
-@user_bp.route("/")
+@bp.route("/")
 @client.require_session()
 async def index() -> str:
     user_info = await client.get_user_info()
     return await render_template("user_home.html", user_info=user_info)
 
 
-@user_bp.route('/passwd', methods=["GET", "POST"])
+@bp.route('/passwd', methods=["GET", "POST"])
 @client.require_session()
 async def change_pw() -> typing.Union[str, quart.Response]:
     form = ChangePasswordForm()
@@ -81,7 +83,7 @@ async def change_pw() -> typing.Union[str, quart.Response]:
     return await render_template("user_passwd.html", form=form)
 
 
-@user_bp.route("/profile", methods=["GET", "POST"])
+@bp.route("/profile", methods=["GET", "POST"])
 @client.require_session()
 async def profile() -> typing.Union[str, quart.Response]:
     form = ProfileForm()
@@ -106,12 +108,12 @@ async def profile() -> typing.Union[str, quart.Response]:
     return await render_template("user_profile.html", form=form)
 
 
-@user_bp.route("/logout", methods=["GET", "POST"])
+@bp.route("/logout", methods=["GET", "POST"])
 @client.require_session()
 async def logout() -> typing.Union[quart.Response, str]:
     form = LogoutForm()
     if form.validate_on_submit():
         await client.logout()
-        return redirect(url_for("home"))
+        return redirect(url_for("main.home"))
 
     return await render_template("user_logout.html", form=form)
